pyproject.toml
[tool.poetry]
name = "muesliswap-onchain-governance"
version = "0.1.0"
description = ""
authors = ["MuesliSwapTeam <contact@muesliswap.com>"]
readme = "README.md"
packages = [{include = "muesliswap_onchain_governance"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
opshin = "0.21.1"
fire = "^0.6.0"
pycardano = {git = "https://github.com/OpShin/pycardano", rev = "running_fixes"}
peewee = "^3.17.3"
ogmios = "^1.0.6"
fastapi = "^0.110.1"
uvicorn = {extras = ["all"], version = "^0.29.0"}
gelidum = "^0.7.0"
fastapi-cache2 = "^0.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
pytest = "^7.4.3"
pre-commit = "^3.5.0"
hypothesis = "^6.92.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
